Consolidate Single enum for Maestro Platform
