--- maestro-client/src/main/java/maestro/Platform.kt
@@ -1,5 +1,0 @@
-package maestro
-
-enum class Platform {
-    ANDROID, IOS, WEB
-}
