+++ maestro-client/src/test/java/maestro/utils/LocaleUtilsTest.kt
@@ -1,6 +1,6 @@
 package maestro.utils
 
-import maestro.Platform
+import maestro.device.Platform
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import com.google.common.truth.Truth.assertThat
