+++ maestro-orchestra/src/main/java/maestro/orchestra/yaml/YamlCondition.kt
@@ -1,7 +1,7 @@
 package maestro.orchestra.yaml
 
 import com.fasterxml.jackson.annotation.JsonFormat
-import maestro.Platform
+import maestro.device.Platform
 
 data class YamlCondition(
     @JsonFormat(with = [JsonFormat.Feature.ACCEPT_CASE_INSENSITIVE_PROPERTIES])
