+++ maestro-studio/server/src/main/java/maestro/studio/Models.kt
@@ -1,7 +1,7 @@
 package maestro.studio
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import maestro.Platform
+import maestro.device.Platform
 import java.util.UUID
 
 data class DeviceScreen(
