+++ maestro-test/src/main/kotlin/maestro/test/drivers/FakeDriver.kt
@@ -28,7 +28,7 @@ import maestro.Driver
 import maestro.KeyCode
 import maestro.MaestroException
 import maestro.OnDeviceElementQuery
-import maestro.Platform
+import maestro.device.Platform
 import maestro.Point
 import maestro.ScreenRecording
 import maestro.SwipeDirection
